requirements.txt
botocore==1.20.74
colorama==0.4.3
docutils==0.15.2
jmespath==0.10.0
Pillow==11.0.0
pyasn1==0.4.8
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reportlab==4.2.5
rsa==4.7.2
s3transfer==0.4.2
six==1.12.0
urllib3==1.26.4
numpy==2.1.2
matplotlib==3.9.2
